Format Option 2 in README.md as code block
The previous formatting mixed code not formatted as code, code formatted as code, and code that was nearly formatted as code in an ordered list. This unnecessarily complicated things and made it more difficult to simply copy and paste the commands.

This commit re-formats the "Option 2" section of the README to make it easier to read and easier to use.